requirements.txt
asgiref==3.4.1
azure-core==1.7.0
azure-functions
azure-storage-blob==12.4.0
beautifulsoup4==4.10.0
blis==0.7.4
cached-property==1.5.2
catalogue==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cymem==2.0.5
deep-translator==1.5.4
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
joblib==1.0.1
more-itertools==8.10.0
murmurhash==1.0.5
nltk==3.6.3
numpy==1.21.3
ordered-set==3.1.1
ordered-set-stubs==0.1.3
pandas==1.3.3
pip==21.2.2
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.5
profanity-filter==1.3.3
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
regex==2021.9.30
requests==2.26.0
river==0.8.0
ruamel.yaml==0.15.100
scikit-learn==1.0.0
setuptools==58.0.4
six==1.16.0
soupsieve==2.2.1
spacy==2.3.7
#spacy<3.1.0,>=2.0
#https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz#egg=en
#https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.1.0/es_core_news_sm-3.1.0.tar.gz#egg=es_core_news_sm
#https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.1.0/zh_core_web_sm-3.1.0.tar.gz#egg=zh_core_web_sm
srsly==1.0.5
starlette==0.14.2
textblob==0.15.3
thinc==7.4.5
tomlkit==0.5.11
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
wasabi==0.8.2
watchgod==0.7
websockets==10.0
wheel==0.37.0
